# server/docs/swagger.yaml
basePath: /api
...
info:
  contact: {}
  description: online exam APIs
  title: online-exam API
  version: 0.0.1
paths:
...
  /oauth/ourspark:
    get:
      parameters:
      - description: code
        in: query
        name: code
        required: true
        type: string
      summary: OAuth 登录
      tags:
      - OAuth
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"